[CL-2008] Extract stacked bars
