+++ front/app/modules/commercial/analytics/admin/components/PostFeedbackCard/StackedBars/utils.ts
